--- sematic/examples/lightning_cifar100_classifier/README.md
@@ -1,6 +1,0 @@
-An example of using
-[PyTorch Lightning](https://pytorch-lightning.readthedocs.io/en/stable/index.html)
-to do distributed training of a model classifying images in the
-[CIFAR100 dataset](https://www.cs.toronto.edu/~kriz/cifar.html).
-
-This example can only be run on Kubernetes clusters set up with specific Kubernetes node types available, and is primarily useful as a reference.
