+++ sematic/examples/lightning_resnet/AUTHORS
