+++ sematic/examples/lightning_resnet/BUILD
